daoServices/flask-appbuilder-based/requirements.txt
#Web Server
flask

#DB Models, SQLAlchemy for flask basically
flask_sqlalchemy

#PostgreSQL DB Connection Adapter
psycopg2-binary

#REST API Generator
flask-appbuilder